requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1
djangorestframework==3.13.1
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.1
Pillow==9.2.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
translate==3.6.1
urllib3==1.26.12
